+++ actions/chats.js
@@ -1,6 +1,6 @@
 'use strict';
 
-import Chat from '../models/Chat';
+import { notifyMessage } from '../utils/notification';
 import getSocket from '../pages/socket';
 
 import types from './types';
@@ -55,10 +55,16 @@ export const receiveChat = chat => dispatch => {
 
 export const receiveMessage = (chatId, message) => (dispatch, getState) => {
     const { user, chats } = getState();
-    const chat = new Chat(chats.find(c => c._id === chatId));
+    const chat = chats.find(c => c._id === chatId);
 
     dispatch({ type: types.RECEIVE_MESSAGE, chatId, message });
-    notifyMessage({ message, chat, user, dispatch });
+
+    const onclick = () => {
+        window.focus();
+        dispatch(openChat(chat._id));
+    };
+
+    notifyMessage({ message, chat, user, dispatch, onclick });
 };
 
 export const updateMessage = (chatId, message) => dispatch => {
@@ -72,42 +78,3 @@ export const updateChat = chat => dispatch => {
 export const destructMessage = (chatId, messageId) => dispatch => {
     dispatch({ type: types.DESTRUCT_MESSAGE, chatId, messageId });
 };
-
-function notifyMessage({ message, chat, user, dispatch }) {
-    const chatInfo = chat.type === 'group' ? ` in ${chat.title}` : '';
-    const title = `@${message.author}${chatInfo}`;
-    const body = message.forwardFrom
-        ? message.forwardFrom.text
-        : message.text;
-    const icon = chat.getAvatarFor(user).replace('.svg', '.png');
-
-    const onclick = () => {
-        window.focus();
-        dispatch(openChat(chat._id));
-    };
-
-    notify({ title, body, icon, onclick });
-}
-
-function notify({ title, body, icon, sound, onclick, vibrate, checkVisibility = true }) {
-    if (checkVisibility && !document.hidden) {
-        return;
-    }
-
-    if (!('Notification' in window)) {
-        return;
-    }
-
-    Notification.requestPermission(permission => {
-        if (permission === 'granted') {
-            const notification = new Notification(title, {
-                icon,
-                body,
-                sound,
-                vibrate
-            });
-
-            notification.onclick = () => onclick();
-        }
-    });
-}
