+++ utils/.eslintrc
@@ -1,0 +1,10 @@
+{
+    "extends": "iddqd/react",
+    "parser": "babel-eslint",
+    "rules": {
+        "capitalized-comments": 0
+    },
+    "env": {
+        "browser": true
+    }
+}
